Implementing new model loading strategy
Doryta's previous loading mechanism assumed that the only neuron network
architecture style that could be build was a variation of building
NNs by layers. This new mechanism allows for arbitrary architectures.

Basically, with this strategy we must specify a graph and its weights.
This is very wasteful for networks with repeated synapse parameters like
convolutional connections. This mode should allow for the highest
flexibility possible.